\name{cdfBeta}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Beta CDF for any x, based on the two shape parameters sh1 and sh2 given in parameters of cdfBeta.}
\usage{
cdfBeta(sh1, sh2)
}
\alias{cdfBeta}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Beta CDF for any x, based on the two shape parameters sh1 and sh2 given in parameters of cdfBeta.
}
\arguments{
  \item{sh1}{shape parameter 1 of the Beta CDF.}
  \item{sh2}{shape parameter 2 of the Beta CDF.}
}
\value{
Returns a function calculating the Gaussian CDF for any x, based on the two shape parameters sh1 and sh2 given in parameters of cdfBeta.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}
