\name{cdfGamma}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Gamma CDF for any x, based on the shape (sh), and scale (sc) given in parameters of cdfGamma.}
\usage{
cdfGamma(sh, sc)
}
\alias{cdfGamma}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Gamma CDF for any x, based on the shape (sh), and scale (sc) given in parameters of cdfGamma.
}
\arguments{
  \item{sh}{Shape parameter of the Gamma CDF.}
  \item{sc}{Scale parameter of the Gamma CDF.}
}
\value{
Returns a function calculating the Gamma CDF for any x, based on the shape (sh), and scale (sc) given in parameters of cdfGamma.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

