\name{cdfGaussian}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Gaussian CDF for any x, based on the mean (m) and standard deviation (ect) given in parameters of cdfGaussian.}
\usage{
cdfGaussian(m, ect)
}
\alias{cdfGaussian}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Gaussian CDF for any x, based on the mean (m) and standard deviation (ect) given in parameters of cdfGaussian.
}
\arguments{
  \item{m}{Mean parameter of the Gaussian CDF.}
  \item{ect}{Standard deviation of the Gaussian CDF.}
}
\value{
Returns a function calculating the Gaussian CDF for any x, based on the mean (m) and standard deviation (ect) given in parameters of cdfGaussian.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}
