\name{cdfGPD}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the GPD CDF for any x, based on the location (mu), shape (xi), and scale (sc) given in parameters of cdfGPD.}
\usage{
cdfGPD(mu, xi, sc)
}
\alias{cdfGPD}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the GPD CDF for any x, based on the location (mu), shape (xi), and scale (sc) given in parameters of cdfGPD.
}
\arguments{
  \item{mu}{Location parameter of the GPD CDF.}
  \item{xi}{Shape parameter of the GPD CDF.}
  \item{sc}{Scale parameter of the GPD CDF.}
}
\value{
Returns a function calculating the GPD CDF for any x, based on the location (mu), shape (xi), and scale (sc) given in parameters of cdfGPD.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}


