\name{CDFt.Beta}
\title{Downscaling or bias correction of Beta CDF through CDF-transformation}
\usage{
CDFt.Beta(ObsRp, DataGp, DataGf, npas = 100, sh1.ini = 0.2, sh2.ini = 0.2)
}
\alias{CDFt.Beta}
\description{
  Downscales (or corrects the model outputs) Beta cumulative distribution function (CDF) of a climate variable from large- to local-scale by applying an equivalent of proportionality transformation :
  i.e., based on a Beta CDF representing a variable at large scale (i.e., low spatial resolution) and the equivalent Beta CDF at a local scale (e.g., modeled at a weather station), this method finds a mathematical transformation allowing to go from the large- to the local-scale CDF. Hence, when a new large-scale Beta CDF is given, a new local-scale CDF (not necessarily Beta) is downscaled based on this transformation.
}
\arguments{
  \item{ObsRp}{Observed time series of the variable (e.g., temperature) at the local scale to be used for estimation of the calibration local-scale Beta CDF.}
  \item{DataGp}{Large-scale time series to be used for estimation of the calibration large-scale Beta CDF.}
  \item{DataGf}{Large-scale time series to be used for estimation of the large-scale Beta CDF to be downscaled.}
  \item{npas}{Number of "cuts" for which quantiles will be estimated (Default is 100).}
  \item{sh1.ini}{Initial value of the shape1 parameter of the beta CDFs (default is 0.2).}
  \item{sh2.ini}{Initial value of the shape2 parameter of the beta CDFs (default is 0.2).}
}
\value{
  CDFt.Beta returns a list with components
  \item{DS}{Downscaled time series generated by "Quantile-matching" method performed between large-scale Beta CDF to be downscaled, and the local-scale downscaled CDF. Note that the length of this array is equal to the length of DataGf}
  \item{x}{an array containing values of the variable (e.g., relative humidity) where the downscaled (and other) CDF has been estimated. WARNING: to be coherent with the Beta CDF, the data have to be in ]0,1[.}
  \item{FRp}{an array containing the values of the local-scale Beta CDF used for calibration, evaluated at the points in x.}
  \item{FGp}{an array containing the values of the large-scale Beta CDF used for calibration, evaluated at the points in x.}
  \item{FGf}{an array containing the values of the large-scale Beta CDF used for downscaling, evaluated at the points in x.}
  \item{FRf}{an array containing the values of the downscaled CDF evaluated at the points in x.}
}
\details{This function provides a parametric (Beta) version of the CDFt function. For details on the (empirical) mathematical formulation of the transformation used to translate the large-scale CDF to the local-scale one, see the reference below.

P.-A. Michelangeli, M. Vrac, H. Loukos. "Probabilistic downscaling approaches: Application to wind cumulative distribution functions", Geophys. Res. Lett.,  doi:10.1029/2009GL038401, 2009.
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr)}
\seealso{\code{\link{CDFt}}, \code{\link{CDFt.Gaussian}}, \code{\link{CDFt.Weibull}}, \code{\link{CDFt.GPD}}, \code{\link{CDFt.Gamma.Dirac}}}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
