\name{CDFt.Gamma.Dirac}
\title{Downscaling or bias correction of Gamma.Dirac CDF (i.e., a mixture of a Gamma distribution and a point mass in 0) through CDF-transformation}
\usage{
CDFt.Gamma.Dirac(ObsRp, DataGp, DataGf, npas = 100, dev = 2,
                    ini.sh.O=NA, ini.sc.O=NA,
                    ini.sh.Gp=NA, ini.sc.Gp=NA,
                    ini.sh.Gf=NA, ini.sc.Gf=NA,
                    ini.sh.GpC=NA, ini.sc.GpC=NA,
                    ini.sh.GfC=NA, ini.sc.GfC=NA)
}
\alias{CDFt.Gamma.Dirac}
\description{
  Downscales (or corrects the model outputs) Gamma.Dirac cumulative distribution function (CDF) of a climate variable from large- to local-scale by applying a equivalent of proportionality transformation :
  i.e., based on a Gamma.Dirac CDF representing a variable at large scale (i.e., low spatial resolution) and the equivalent Gamma.Dirac CDF at a local scale (e.g., modeled at a weather station), this method finds a mathematical transformation allowing to go from the large- to the local-scale CDF. Hence, when a new large-scale Gamma.Dirac CDF is given, a new local-scale CDF (not necessarily Gamma.Dirac) is downscaled based on this transformation.
}
\arguments{
  \item{ObsRp}{Observed time series of the variable (e.g., temperature) at the local scale to be used for estimation of the calibration local-scale Gamma.Dirac CDF.}
  \item{DataGp}{Large-scale time series to be used for estimation of the calibration large-scale Gamma.Dirac CDF.}
  \item{DataGf}{Large-scale time series to be used for estimation of the large-scale Gamma.Dirac CDF to be downscaled.}
  \item{npas}{Number of "cuts" for which quantiles will be estimated (Default is 100).}
  \item{dev}{Coefficient of development (of the difference between the mean of the large-scale historical data and the mean of the large-scale future data to be downscaled). This development is used to extend range of data on which the quantiles will be calculated for the CDF to be downscaled (Default is 2).}
  \item{ini.sh.O}{Initialisation value for the Gamma CDF shape parameter of the observed (i.e., present local-scale) data.}
  \item{ini.sc.O}{Same as ini.sh.O but for the scale parameter.}
  \item{ini.sh.Gp}{Same as ini.sh.O but for the large-scale present data.}
  \item{ini.sc.Gp}{Same as ini.sh.Gp but for the scale parameter.}
  \item{ini.sh.Gf}{Same as ini.sh.Gp but for the large-scale future data.}
  \item{ini.sc.Gf}{Same as ini.sh.Gf but for the scale parameter.}
  \item{ini.sh.GpC}{Same as ini.sh.Gp for the "centered" present large-scale data.}
  \item{ini.sc.GpC}{Same as ini.sh.GpC but for the scale parameter.}
  \item{ini.sh.GfC}{Same as ini.sh.GpC but for future "centered" large-scale data.}
  \item{ini.sc.GfC}{Same as ini.sh.GfC but for the scale parameter.}
}
\value{
  CDFt.Gamma.Dirac returns a list with components
  \item{DS}{Downscaled time series generated by "Quantile-matching" method performed between large-scale Gamma.Dirac CDF to be downscaled, and the local-scale downscaled CDF. Note that the length of this array is equal to the length of DataGf}
  \item{x}{an array containing values of the variable (e.g., temperature) where the downscaled (and other) CDF has been estimated.}
  \item{FRp}{an array containing the values of the local-scale Gamma.Dirac CDF used for calibration, evaluated at the points in x.}
  \item{FGp}{an array containing the values of the large-scale Gamma.Dirac CDF used for calibration, evaluated at the points in x.}
  \item{FGf}{an array containing the values of the large-scale Gamma.Dirac CDF used for downscaling, evaluated at the points in x.}
  \item{FRf}{an array containing the values of the downscaled CDF evaluated at the points in x.}
  \item{p0Rf}{Estimated probability of the point mass in 0.}
}
\details{This function provides a parametric (Gamma.Dirac) version of the CDFt function. For details on the (empirical) mathematical formulation of the transformation used to translate the large-scale CDF to the local-scale one, see the reference below. Note that in this R package, the large-scale data (i.e., DataGp and DataGf) are automatically transformed to have the same mean as the ObsRp time series. This avoid to get out of the range of applicability of the CDFt method. However, the large-scale output CDFs have their initial mean (i.e., not centered).

Note that if the parameters ini.sh.O... ini.sc.GfC are set to NaN (i.e., the default), the initialization of the shape and scale parameters is performed by the method of moments.

P.-A. Michelangeli, M. Vrac, H. Loukos. "Probabilistic downscaling approaches: Application to wind cumulative distribution functions", Geophys. Res. Lett.,  doi:10.1029/2009GL038401, 2009.
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr)}
\seealso{\code{\link{CDFt}}, \code{\link{CDFt.Gaussian}}, \code{\link{CDFt.GPD}}, \code{\link{CDFt.Weibull}}}
\examples{
## Example

### Generation of example data
O = rGammaDirac(n=1500, sh=1, sc=0.5, p0=0.2)
Gp = rGammaDirac(n=1500, sh=0.7, sc=0.6, p0=0.3)
Gf = rGammaDirac(n=1500, sh=1, sc=0.5, p0=0.3)

### Call of the CDFt.Gamma.Dirac method
CG = CDFt.Gamma.Dirac(O,Gp,Gf)

x <- CG$x
FGp <- CG$FGp
FGf <- CG$FGf
FRp <- CG$FRp
FRf <- CG$FRf
ds <- CG$DS
p0Rf = CG$p0Rf


### Plot the results
par(mfrow=c(1,2))

plot(x, FGp,type="l",lty=2,ylim=c(0,1),xlab="x",ylab="CDF(x)")
lines(x,FGf,type="l",lty=2,col=2)
lines(x,FRp,type="l")
lines(x,FRf,type="l",col=2)

# Compare FRf with empirical CDF from downscaled values
e = ecdf(ds)
lines(x,e(x),type="l",lty=2,col=4)

plot(Gf,ds,xlab="Large-scale data", ylab="Downscaled data")

}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
