\name{CDFt.Gaussian}
\title{Downscaling or bias correction of Gaussian CDF through CDF-transformation}
\usage{
CDFt.Gaussian(ObsRp, DataGp, DataGf, npas = 100, dev = 2)
}
\alias{CDFt.Gaussian}
\description{
  Downscales (or corrects the model outputs) Gaussian cumulative distribution function (CDF) of a climate variable from large- to local-scale by applying a equivalent of proportionality transformation :
  i.e., based on a Gaussian CDF representing a variable at large scale (i.e., low spatial resolution) and the equivalent Gaussian CDF at a local scale (e.g., modeled at a weather station), this method finds a mathematical transformation allowing to go from the large- to the local-scale CDF. Hence, when a new large-scale Gaussian CDF is given, a new local-scale CDF (not necessarily Gaussian) is downscaled based on this transformation.
}
\arguments{
  \item{ObsRp}{Observed time series of the variable (e.g., temperature) at the local scale to be used for estimation of the calibration local-scale Gaussian CDF.}
  \item{DataGp}{Large-scale time series to be used for estimation of the calibration large-scale Gaussian CDF.}
  \item{DataGf}{Large-scale time series to be used for estimation of the large-scale Gaussian CDF to be downscaled.}
  \item{npas}{Number of "cuts" for which quantiles will be estimated (Default is 100).}
  \item{dev}{Coefficient of development (of the difference between the mean of the large-scale historical data and the mean of the large-scale future data to be downscaled). This development is used to extend range of data on which the quantiles will be calculated for the CDF to be downscaled (Default is 2).}
}
\value{
  CDFt.Gaussian returns a list with components
  \item{DS}{Downscaled time series generated by "Quantile-matching" method performed between large-scale Gaussian CDF to be downscaled, and the local-scale downscaled CDF. Note that the length of this array is equal to the length of DataGf}
  \item{x}{an array containing values of the variable (e.g., temperature) where the downscaled (and other) CDF has been estimated.}
  \item{FRp}{an array containing the values of the local-scale Gaussian CDF used for calibration, evaluated at the points in x.}
  \item{FGp}{an array containing the values of the large-scale Gaussian CDF used for calibration, evaluated at the points in x.}
  \item{FGf}{an array containing the values of the large-scale Gaussian CDF used for downscaling, evaluated at the points in x.}
  \item{FRf}{an array containing the values of the downscaled CDF evaluated at the points in x.}
}
\details{This function provides a parametric (Gaussian) version of the CDFt function. For details on the (empirical) mathematical formulation of the transformation used to translate the large-scale CDF to the local-scale one, see the reference below. Note that in this R package, the large-scale data (i.e., DataGp and DataGf) are automatically transformed to have the same mean as the ObsRp time series. This avoid to get out of the range of applicability of the CDFt method. However, the large-scale output CDFs have their initial mean (i.e., not centered).

P.-A. Michelangeli, M. Vrac, H. Loukos. "Probabilistic downscaling approaches: Application to wind cumulative distribution functions", Geophys. Res. Lett.,  doi:10.1029/2009GL038401, 2009.
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr)}
\seealso{\code{\link{CDFt}}, \code{\link{CDFt.Weibull}}, \code{\link{CDFt.GPD}}, \code{\link{CDFt.Gamma.Dirac}}}
\examples{
## Example

### Generation of example data
O <- rnorm(2100,mean=0,sd=1)
Gp <- rnorm(300,mean=3,sd=1)
Gf <- rnorm(300,mean=4,sd=1)

### Call of the CDFt.Gaussian method
CG = CDFt.Gaussian(O,Gp,Gf)

x <- CG$x
FGp <- CG$FGp
FGf <- CG$FGf
FRp <- CG$FRp
FRf <- CG$FRf
ds <- CG$DS

### Plot the results
par(mfrow=c(1,2))

plot(x, FGp,type="l",lty=2,ylim=c(0,1),xlab="x",ylab="CDF(x)")
lines(x,FGf,type="l",lty=2,col=2)
lines(x,FRp,type="l")
lines(x,FRf,type="l",col=2)

# Compare FRf with empirical CDF from downscaled values
e = ecdf(ds)
lines(x,e(x),type="l",lty=2,col=4)

plot(Gf,ds,xlab="Large-scale data", ylab="Downscaled data")

}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
