\name{CDFt.threshold.Gamma}
\title{Downscaling or bias correction of Gamma CDF through CDF-transformation after thresholding}
\usage{
CDFt.threshold.Gamma(ObsRp, DataGp, DataGf, thObs = 1,
                    npas = 100, dev = 2,
                    ini.sh.O=NA, ini.sc.O=NA,
                    ini.sh.Gp=NA, ini.sc.Gp=NA,
                    ini.sh.Gf=NA, ini.sc.Gf=NA)
}
\alias{CDFt.threshold.Gamma}
\description{
  Downscales (or corrects the model outputs) Gamma cumulative distribution function (CDF) of a climate variable from large- to local-scale by applying a equivalent of proportionality transformation :
  i.e., based on a Gamma CDF representing a variable at large scale (i.e., low spatial resolution) and the equivalent Gamma CDF at a local scale (e.g., modeled at a weather station), this method finds a mathematical transformation allowing to go from the large- to the local-scale CDF. Hence, when a new large-scale Gamma CDF is given, a new local-scale CDF (not necessarily Gamma) is downscaled based on this transformation.
Before applying this transformation, a threshold ding is applied on all datasets (see details).
WARNING: It is assumed that the input data are >0 to compute Gamma CDFs. Thus, the obs threshold (thObs) must be >0.
}
\arguments{
  \item{ObsRp}{Observed time series of the variable (e.g., temperature) at the local scale to be used for estimation of the calibration local-scale Gamma CDF.}
  \item{DataGp}{Large-scale time series to be used for estimation of the calibration large-scale Gamma CDF.}
  \item{DataGf}{Large-scale time series to be used for estimation of the large-scale Gamma CDF to be downscaled.}
  \item{thObs}{Threshold chosen for the observation data. Below this threshold (default = 1), values are considered as 0.}
  \item{npas}{Number of "cuts" for which quantiles will be estimated (Default is 100).}
  \item{dev}{Coefficient of development (of the difference between the mean of the large-scale historical data and the mean of the large-scale future data to be downscaled). This development is used to extend range of data on which the quantiles will be calculated for the CDF to be downscaled (Default is 2).}
  \item{ini.sh.O}{Initialisation value for the Gamma CDF shape parameter of the observed (i.e., present local-scale) data.}
  \item{ini.sc.O}{Same as ini.sh.O but for the scale parameter.}
  \item{ini.sh.Gp}{Same as ini.sh.O but for the large-scale present data.}
  \item{ini.sc.Gp}{Same as ini.sh.Gp but for the scale parameter.}
  \item{ini.sh.Gf}{Same as ini.sh.Gp but for the large-scale future data.}
  \item{ini.sc.Gf}{Same as ini.sh.Gf but for the scale parameter.}
}
\value{
  CDFt.Gamma returns a list with components
  \item{DS}{Downscaled time series generated by "Quantile-matching" method performed between large-scale Gamma CDF to be downscaled, and the local-scale downscaled CDF. Note that the length of this array is equal to the length of DataGf}
  \item{x}{an array containing values of the variable (e.g., temperature) where the downscaled (and other) CDF has been estimated.}
  \item{FRp}{an array containing the values of the local-scale Gamma CDF used for calibration, evaluated at the points in x.}
  \item{FGp}{an array containing the values of the large-scale Gamma CDF used for calibration, evaluated at the points in x.}
  \item{FGf}{an array containing the values of the large-scale Gamma CDF used for downscaling, evaluated at the points in x.}
  \item{FRf}{an array containing the values of the downscaled CDF evaluated at the points in x.}
}
\details{This function provides a parametric (Gamma) version of the CDFt function.
The large-scale threshold thG is chosen with respect to obs threshold thObs such that Pobs(obs<=thObs)~=Pgcm(GCM<=tG) by QQ-matching. This is done on the DataGp (training) set.
For details on the (empirical) mathematical formulation of the transformation used to translate the large-scale CDF to the local-scale one, see the reference below.

Note that if the parameters ini.sh.O... are set to NaN (i.e., the default), the initialization of the shape and scale parameters is performed by the method of moments.

P.-A. Michelangeli, M. Vrac, H. Loukos. "Probabilistic downscaling approaches: Application to wind cumulative distribution functions", Geophys. Res. Lett.,  doi:10.1029/2009GL038401, 2009.
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr)}
\seealso{\code{\link{CDFt}}, \code{\link{CDFt.Gaussian}}, \code{\link{CDFt.GPD}}, \code{\link{CDFt.Gamma}}, \code{\link{CDFt.Weibull}}}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
