\name{cdfWeibull}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Weibull CDF for any x, based on the shape (sh) and scale (sc) given in parameters of cdfWeibull.}
\usage{
cdfWeibull(sh, sc)
}
\alias{cdfWeibull}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the Weibull CDF for any x, based on the shape (sh) and scale (sc) given in parameters of cdfWeibull.
}
\arguments{
  \item{sh}{Shape parameter of the Weibull CDF.}
  \item{sc}{Scale parameter of the Weibull CDF.}
}
\value{
Returns a function calculating the Weibull CDF for any x, based on the shape (sh) and scale (sc) given in parameters of cdfWeibull.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

