\name{MinusLogLikBeta}
\title{This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Beta distribution from data.}
\usage{
MinusLogLikBeta(X, sh1, sh2)
}
\alias{MinusLogLikBeta}
\description{
This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Beta distribution from data.
}
\arguments{
  \item{X}{Vector of data to compute the log-likelihood}
  \item{sh1}{Shape parameter 1 of the Beta CDF.}
  \item{sh2}{Shape parameter 2 of the Beta CDF.}
}
\value{
Returns minus (-) log-likelihood of a Beta distribution from data.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}


