\name{MinusLogLikGamma}
\title{This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Gamma distribution from data.}
\usage{
MinusLogLikGamma(X, lambda, gam)
}
\alias{MinusLogLikGamma}
\description{
This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Gamma distribution from data.
}
\arguments{
  \item{X}{Vector of data.}
  \item{lambda}{Shape parameter of the Gamma CDF.}
  \item{gam}{Rate (=1/scale) parameter of the Gamma CDF.}
}
\value{
Returns minus (-) log-likelihood of a Gamma distribution from data.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

