\name{MinusLogLikWeibull}
\title{This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Weibull distribution from data.}
\usage{
MinusLogLikWeibull(X, shape, scale)
}
\alias{MinusLogLikWeibull}
\description{
This function is for internal use and is not meant to be called by
the user. It provides the minus (-) log-likelihood of a Weibull distribution from data.
}
\arguments{
  \item{X}{Vector of data to compute the log-likelihood}
  \item{shape}{Shape parameter of the Weibull CDF.}
  \item{scale}{Scale parameter of the Weibull CDF.}
}
\value{
Returns minus (-) log-likelihood of a Weibull distribution from data.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}


