\name{Non.neg.values1D}
\title{This function is for internal use and is not meant to be called by
the user. It returns >0 values from a 1d-vector.}
\usage{
Non.neg.values1D(Tab1d)
}
\alias{Non.neg.values1D}
\description{
This function is for internal use and is not meant to be called by
the user. It returns >0 values from a 1d-vector.
}
\arguments{
  \item{tab1d}{A vector of data.}
}
\value{
Returns >0 values from a 1d-vector.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}


