\name{pgpd.mk}
\title{This function is for internal use and is not meant to be called by the user.
}
\usage{
pgpd.mk(q, xi, mu, sc)
}
\alias{pgpd.mk}
\description{
This function is for internal use and is not meant to be called by
the user.
It computes the cumulative distribution value (F(q)) of a quantile value q, from a GPD CDF with parameters (xi, mu,sc).
}
\arguments{
  \item{q}{Quantile value.}
  \item{xi}{Shape parameter of the GPD CDF.}
  \item{mu}{Location parameter of the GPD CDF.}
  \item{sc}{Scale parameter of the GPD CDF.}
}
\value{
computes the cumulative distribution value (F(q)) of a quantile value q, from a GPD CDF with parameters (xi, mu,sc).
}
\details{This function is for internal use and is not meant to be called by the user.}
\keyword{internal}


