\name{quantileBeta}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Beta CDF for any probability p, based on the two shape partameters sh1 and sh2 given in parameters of quantileBeta.}
\usage{
quantileBeta(sh1, sh2)
}
\alias{quantileBeta}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Beta CDF for any probability p, based on the two shape partameters sh1 and sh2 given in parameters of quantileBeta.
}
\arguments{
  \item{sh1}{shape parameter 1 of the Beta CDF.}
  \item{sh2}{shape parameter 2 of the Beta CDF.}
}
\value{
Returns a function calculating the inverse Beta CDF for any probability p, based on the two shape partameters sh1 and sh2 given in parameters of quantileBeta.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

