\name{quantileGamma}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Gamma CDF for any probability p, based on the shape (sh), and scale (sc) given in parameters of quantileGamma.}
\usage{
quantileGamma(sh, sc)
}
\alias{quantileGamma}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Gamma CDF for any probability p, based on the shape (sh), and scale (sc) given in parameters of quantileGamma.
}
\arguments{
  \item{sh}{Shape parameter of the Gamma CDF.}
  \item{sc}{Scale parameter of the Gamma CDF.}
}
\value{
Returns a function calculating the inverse Gamma CDF for any probability p, based on the shape (sh), and scale (sc) given in parameters of quantileGamma.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

