\name{quantileGaussian}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Gaussian CDF for any probability p, based on the mean (m) and standard deviation (ect) given in parameters of quantileGaussian.}
\usage{
quantileGaussian(m, ect)
}
\alias{quantileGaussian}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Gaussian CDF for any probability p, based on the mean (m) and standard deviation (ect) given in parameters of quantileGaussian.
}
\arguments{
  \item{m}{Mean parameter of the Gaussian CDF.}
  \item{ect}{Standard deviation of the Gaussian CDF.}
}
\value{
Returns a function calculating the inverse Gaussian CDF for any probability p, based on the mean (m) and standard deviation (ect) given in parameters of quantileGaussian.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

