\name{quantileWeibull}
\title{This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Weibull CDF for any probability p, based on the shape (sh) and scale (sc) given in parameters of quantileWeibull.}
\usage{
quantileWeibull(sh, sc)
}
\alias{quantileWeibull}
\description{
This function is for internal use and is not meant to be called by
the user. It returns a function calculating the inverse Weibull CDF for any probability p, based on the shape (sh) and scale (sc) given in parameters of quantileWeibull.
}
\arguments{
  \item{sh}{Shape parameter of the Weibull CDF.}
  \item{sc}{Scale parameter of the Weibull CDF.}
}
\value{
Returns a function calculating the inverse Weibull CDF for any probability p, based on the shape (sh) and scale (sc) given in parameters of quantileWeibull.
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

