\name{rGammaDirac}
\title{This function is for internal use and is not meant to be called by
the user. It randomly generates values according to a CDF corresponding to a mixture of a Gamma distribution and a point mass in 0 (i.e., Dirac).}
\usage{
rGammaDirac(n, sh, sc, p0)
}
\alias{rGammaDirac}
\description{
This function is for internal use and is not meant to be called by
the user. It randomly generates values according to a CDF corresponding to a mixture of a Gamma distribution and a point mass in 0 (i.e., Dirac).
}
\arguments{
  \item{n}{Number of values to be generated.}
  \item{sh}{Shape parameter of the Gamma CDF.}
  \item{sc}{Scale parameter of the Gamma CDF.}
  \item{p0}{Probability of the point mass in 0.}
}
\value{
Randomly generates values according to a CDF corresponding to a mixture of a Gamma distribution and a point mass in 0 (i.e., Dirac).
}
\details{This function is for internal use and is not meant to be called by
the user.}
\keyword{math}
\keyword{distribution}
\keyword{univar}
\keyword{models}
\keyword{internal}

